requirements.txt
Image
wand
pytesseract
pyenchant